setu-aa-fi/requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
dnspython==1.16.0
idna==3.2
pycparser==2.20
PyJWT==2.4.0
pymongo==3.11.2
requests==2.26.0
urllib3==1.26.7
